Allow '+' and '-' to be read as symbols.
After reading '+' or '-' at the start of an expression, look ahead to
the following character to see whether to read a number or the prefix
as its own symbol. If the next character is a delimiter or does not
exist, read a symbol, otherwise continue trying to read a number.

Rename the 'plus' and 'minus' functions in the default environment
to '+' and '-' respectively.