Add let special form.
Change the type of the environment to be represented as a list
of environment frames an allow new frames to be added and removed.

Implement the let special form which evaulates a list of bindings
in the current environment, adds them to a new environment frame
and then evaluates the given body in the modified environment.