Add tokeniser.
Add Token type and function for reading the next token from a string.
The readToken' function accumulates token characters into a PartialToken
value until the end of the token is reached, when a token value is
returned.

Add readTokens function to read a list of tokens from an input string.