Add quote and unquote literals to the reader.
Add partial quote and unquote forms to the parser state. Since both
quote and unquote require a single value to apply to, encountering
a single value could cause multiple nested forms to be completed.
Replace the parser state with a stack of partial items and complete
as many nested items as possible when a complete value is encountered.