Add set! special form.
Add set! special form which allows the value of a binding to be updated.

Move setBinding from the Env namespace to the Eval namespace and change
it to yield an UnboundRef error if the binding being set does not exist
in the current environment.

Change the let* special form to incrementally build the top-level
environment frame instead of creating a new one for each binding. This
allows previously-bound variables in the frame to be modified with
set!