Add support for super method invocations.
Add a new SuperCall constructor for Value which represents a call to
the superclass. This constructor contains the current receiver along
with the declaring class for the executing method. Super calls are
only resolved along the inheritence chain from the current declaring
class and not from the class of the receiver.

Add a MethodDecl type which contains the name, definition and declaring
class for a method. Add a findMethod function which searches the
inheritence chain to find a method definition in the closest
ancestor class.